fix(main/lua-language-server): Check for TERMUX_EXEC__PROC_SELF_EXE
This fixes running under `termux-exec` with updated `targetSdk`.

Brief summary: When running with `targetSdk` to 28 or higher, it's not allowed to execute downloaded files (files not part of the APK in general) directly.

The termux-exec workaround using `LD_PRELOAD` that is currently used in the Google Play build instead executes e.g. `/system/bin/linker64 $PREFIX/bin/binary` instead of `$PREFIX/bin/binary` directly. One consequence of that is that reading `/proc/self/exe` to get the path to the currently running executable does not work, so patches like these (that should be a no-op when not running under bumped `targetSdk` and `termux-exec`) are necessary.

Fixes issue reported in https://github.com/termux/termux-packages/issues/21275